WIP: Default to DebugMono configuration for now
Since I'm working in VS Code on Linux, the C# plugin uses the default
configuration as its guide to loading the DLLs. To make that work, I
need the default configuration to be DebugMono instead of Debug, so that
the DLLs that I build in that configuration will be found by the editor
plugin and I won't get red wavy lines all over my project.