Add NuGet package Newtonsoft.Json
The DataContractJsonSerializer can't cope with tuples or KeyValuePairs,
and we're going to need at least one of those in order to do the FW->LF
direction properly. As a bonus, now that Newtonsoft.Json is in use in
both LfMerge and LfMergeBridge, we'll be able to write slightly simpler
serialization and deserialization code.