WIP: Don't duplicate already-known comments
This doesn't work, because DataContractJsonSerializer doesn't know how
to handle tuples. Next commit we'll switch to Newtonsoft.Json and, as a
bonus, be able to slightly simplify the LF->FW direction (where we'll
also be able to use tuples or KeyValuePairs, and won't have to send two
separate input files).